requirements.txt
beautifulsoup4==4.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==0.12.5
get==2019.4.13
google-api-core==1.19.1
google-api-python-client==1.9.1
googleapis-common-protos==1.52.0
google-auth==1.16.1
google-auth-httplib2==0.0.3
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pip==19.0.3
post==2019.4.13
protobuf==3.12.2
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
rsa==4.0
setuptools==40.8.0
six==1.15.0
soupsieve==2.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==0.16.1
